feat: implement GEO SEO optimization for AIQUAA Labs
- Add dynamic sitemap with all lab and assessment routes
- Create robots.ts for AI crawler permissions (GPTBot, ClaudeBot, PerplexityBot)
- Restructure lab pages to Server Components with metadata exports
- Add OpenGraph images, Twitter Cards, and canonical URLs for each tool
- Create GEO-optimized content files for key tools (ISTQB, Performance, API Testing, SQL, All Pairs)
- Add JSON-LD schemas for organization, software applications, and rankings
- Create llms.txt for AI crawler guidance
- Add GEO implementation plan documentation

Closes #123
