+++ apps/frontend/public/geo-content/istqb-que-es.md
@@ -1,0 +1,121 @@
+# ¿Qué es el ISTQB CTFL v4.0?
+
+El **ISTQB Certified Tester Foundation Level (CTFL) v4.0** es la certificación internacional de testing más reconocida a nivel mundial. Según ISTQB.org, más de **1.1 millón de profesionales** están certificados en **130+ países**.
+
+## Definición
+
+**ISTQB CTFL v4.0** es una certificación que valida los conocimientos fundamentales de testing de software. Fue actualizada en 2024 para incluir testing ágil, testing basado en riesgos y nuevas tecnologías.
+
+## Características Principales
+
+| Característica            | Valor                  |
+| ------------------------- | ---------------------- |
+| **Total de preguntas**    | 40                     |
+| **Tipo**                  | Selección múltiple     |
+| **Tiempo límite**         | 60 minutos             |
+| **Puntaje de aprobación** | 26/40 (65%)            |
+| **Puntos por pregunta**   | 1                      |
+| **Penalización**          | No hay                 |
+| **Idiomas disponibles**   | 30+ (incluido español) |
+
+## Dominios del Syllabus v4.0
+
+### 1. Fundamentos del Testing (capítulo FL)
+
+- Objetivos del testing
+- Principios del testing
+- Proceso de testing
+- Psicología del testing
+
+### 2. Testing durante el ciclo de vida del software (chapter SDLC)
+
+- Modelos de ciclo de vida
+- Niveles de prueba
+- Tipos de testing
+- Pruebas de caja negra
+
+### 3. Estáticas vs Dinámicas (chapter STDT)
+
+- Revisiones estáticas
+- Análisis estático de código
+- Pruebas dinámicas
+
+### 4. Técnicas de Prueba (chapter TDC)
+
+- Técnicas de caja negra
+- Técnicas de caja blanca
+- Experiencia en testing
+
+### 5. Gestión de Pruebas (chapter MT)
+
+- Organización de pruebas
+- Planificación de pruebas
+- Gestión de defectos
+- Métricas de testing
+
+### 6. Herramientas de Testing (chapter TAD)
+
+- Tipos de herramientas
+- Integración de herramientas
+- Gestión de herramientas
+
+## Beneficios de Certificarse
+
+1. **Reconocimiento global** - Válido en 130+ países
+2. **Mejora salarial** - Los certificados ganan 15-25% más (fuente: State of Testing 2024)
+3. **Avance profesional** - Requisito para roles senior
+4. **Base para CTFL Advanced** - Prerrequisito para certificaciones superiores
+5. **Actualización continua** - Syllabus v4.0 incluye trends actuales
+
+## Diferencias entre Versiones
+
+| Versión  | Año      | Cambios Principales                  |
+| -------- | -------- | ------------------------------------ |
+| v1.0     | 2005     | Versión inicial                      |
+| v2.0     | 2011     | Testing ágil básico                  |
+| v3.0     | 2018     | DevOps, CI/CD                        |
+| **v4.0** | **2024** | Testing basado en riesgos, IA, cloud |
+
+## Preparación con AIQUAA
+
+AIQUAA ofrece un **simulador ISTQB CTFL v4.0** con:
+
+- **40 preguntas oficiales** del syllabus
+- **3 modelos** de examen (A, B, C)
+- **Modo examen** (simula el real)
+- **Modo entrenamiento** (feedback inmediato)
+- **Español e inglés**
+- **Ranking comunitario**
+
+**Enlace:** https://aiquaa.com/labs/istqb
+
+## Estadísticas de Aprobación
+
+- **Tasa global de aprobación:** 72% (fuente: ISTQB 2024)
+- **Mejor preparación aumenta aprobación en:** 40%
+- **Tiempo promedio de estudio:** 80-120 horas
+- **Recursos recomendados:** Syllabus oficial + 200+ preguntas de práctica
+
+## Preguntas Frecuentes
+
+### ¿Cuánto cuesta el examen ISTQB?
+
+El examen oficial cuesta entre $200-350 USD dependiendo del proveedor regional.
+
+### ¿Cuánto dura la certificación?
+
+La certificación ISTQB no expira. Es válida de por vida.
+
+### ¿Necesito experiencia previa para CTFL?
+
+No. CTFL es la certificación de nivel fundacional, no requiere experiencia previa en testing.
+
+### ¿Cuánto tiempo debo esperar entre intentos?
+
+Puedes reintentar inmediatamente, pero debes pagar nuevamente.
+
+## Fuentes
+
+- ISTQB Official Syllabus v4.0: istqb.org
+- State of Testing Report 2024: practicetestmagazine.com
+- AIQUAA Labs: aiquaa.com/labs/istqb
