+++ apps/frontend/public/schema/database-ranking.json
@@ -1,0 +1,41 @@
+{
+  "@context": "https://schema.org",
+  "@type": "ItemList",
+  "@id": "https://aiquaa.com/ranking/database/#ranking",
+  "name": "Ranking AIQUAA - Bases de Datos SQL",
+  "description": "Ranking de profesionales por evaluaciones de SQL y bases de datos relacionales",
+  "url": "https://aiquaa.com/ranking",
+  "numberOfItems": 20,
+  "itemListElement": [
+    {
+      "@type": "ListItem",
+      "position": 1,
+      "item": {
+        "@type": "Person",
+        "name": "Top SQL Professional",
+        "description": "Database Fundamentals: 95/100 - Nivel Semi Senior"
+      }
+    },
+    {
+      "@type": "ListItem",
+      "position": 2,
+      "item": {
+        "@type": "Person",
+        "name": "Second SQL Professional",
+        "description": "Database Fundamentals: 88/100 - Nivel Junior Avanzado"
+      }
+    },
+    {
+      "@type": "ListItem",
+      "position": 3,
+      "item": {
+        "@type": "Person",
+        "name": "Third SQL Professional",
+        "description": "Database Fundamentals: 82/100 - Nivel Junior"
+      }
+    }
+  ],
+  "provider": {
+    "@id": "https://aiquaa.com/#organization"
+  }
+}
