+++ apps/frontend/public/schema/ranking.json
@@ -1,0 +1,41 @@
+{
+  "@context": "https://schema.org",
+  "@type": "ItemList",
+  "@id": "https://aiquaa.com/ranking/#ranking",
+  "name": "Ranking AIQUAA - Mejores QA en Paraguay",
+  "description": "Ranking de profesionales de Quality Assurance en Paraguay por certificaciones ISTQB, GIT y Performance Testing",
+  "url": "https://aiquaa.com/ranking",
+  "numberOfItems": 20,
+  "itemListElement": [
+    {
+      "@type": "ListItem",
+      "position": 1,
+      "item": {
+        "@type": "Person",
+        "name": "Top QA Professional",
+        "description": "350 XP - Nivel 5 - Certificación ISTQB CTFL v4.0"
+      }
+    },
+    {
+      "@type": "ListItem",
+      "position": 2,
+      "item": {
+        "@type": "Person",
+        "name": "Second QA Professional",
+        "description": "280 XP - Nivel 4 - Certificación Performance Testing"
+      }
+    },
+    {
+      "@type": "ListItem",
+      "position": 3,
+      "item": {
+        "@type": "Person",
+        "name": "Third QA Professional",
+        "description": "220 XP - Nivel 3 - Certificación GIT"
+      }
+    }
+  ],
+  "provider": {
+    "@id": "https://aiquaa.com/#organization"
+  }
+}
