+++ apps/frontend/src/app/assessments/database-fundamentals/page.tsx
@@ -1,5 +1,54 @@
+import { Metadata } from 'next';
 import { databaseFundamentalsDefinition } from './data/assessment-definition';
 import AssessmentWelcome from '../_shared/components/AssessmentWelcome';
+
+export const metadata: Metadata = {
+  title: 'Bases de Datos — Fundamentos | AIQUAA',
+  description:
+    'Prueba técnica teórica sobre modelo relacional, SQL básico, joins, agregaciones y constraints. Evalúa tus conocimientos de bases de datos.',
+  keywords: [
+    'bases de datos',
+    'SQL',
+    'modelo relacional',
+    'JOINs',
+    'agregaciones',
+    'constraints',
+    'QA',
+    'testing',
+    'AIQUAA',
+    'evaluación técnica',
+    'fundamentos SQL',
+  ],
+  openGraph: {
+    title: 'Bases de Datos — Fundamentos | AIQUAA',
+    description:
+      'Prueba técnica teórica sobre modelo relacional, SQL básico, joins, agregaciones y constraints.',
+    url: 'https://aiquaa.com/assessments/database-fundamentals',
+    siteName: 'AIQUAA',
+    type: 'website',
+    locale: 'es_PY',
+    images: [
+      {
+        url: '/api/og?title=Bases%20de%20Datos%20-%20Fundamentos&subtitle=Evaluación%20de%20SQL%20y%20modelo%20relacional&section=Assessments',
+        width: 1200,
+        height: 630,
+        alt: 'Bases de Datos Fundamentos - AIQUAA',
+      },
+    ],
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Bases de Datos — Fundamentos | AIQUAA',
+    description: 'Prueba técnica sobre modelo relacional y SQL básico.',
+    images: [
+      '/api/og?title=Bases%20de%20Datos%20-%20Fundamentos&subtitle=Evaluación%20de%20SQL&section=Assessments',
+    ],
+    creator: '@stevenayal',
+  },
+  alternates: {
+    canonical: 'https://aiquaa.com/assessments/database-fundamentals',
+  },
+};
 
 export default function DatabaseFundamentalsPage() {
   const overview = {
