+++ apps/frontend/src/app/robots.ts
@@ -1,0 +1,26 @@
+import type { MetadataRoute } from 'next';
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: [
+      // AI Crawlers — Allow
+      { userAgent: 'GPTBot', allow: '/' },
+      { userAgent: 'OAI-SearchBot', allow: '/' },
+      { userAgent: 'ChatGPT-User', allow: '/' },
+      { userAgent: 'ClaudeBot', allow: '/' },
+      { userAgent: 'anthropic-ai', allow: '/' },
+      { userAgent: 'PerplexityBot', allow: '/' },
+      { userAgent: 'Google-Extended', allow: '/' },
+      { userAgent: 'GoogleOther', allow: '/' },
+      { userAgent: 'Applebot-Extended', allow: '/' },
+      { userAgent: 'Amazonbot', allow: '/' },
+      { userAgent: 'FacebookBot', allow: '/' },
+      // Block aggressive crawlers
+      { userAgent: 'Bytespider', disallow: '/' },
+      // Default
+      { userAgent: '*', allow: '/', disallow: ['/admin/', '/api/'] },
+    ],
+    sitemap: 'https://aiquaa.com/sitemap.xml',
+    host: 'https://aiquaa.com',
+  };
+}
