+++ docs/geo-implementation-plan.md
@@ -1,0 +1,181 @@
+# Plan de Implementación GEO para AIQUAA
+
+## Resumen Ejecutivo
+
+AIQUAA tiene potencial para convertirse en la **referencia en QA para LATAM** mediante optimización para motores de búsqueda AI (GEO). Este plan detalla las acciones para mejorar la visibilidad en ChatGPT, Claude, Perplexity, Gemini y Google AI Overviews.
+
+## Estado Actual
+
+### Herramientas Disponibles
+
+- **23 herramientas** en 8 categorías
+- **7 evaluaciones** con ranking comunitario
+- **Contenido QA** en español
+- **Gamificación** con XP y logros
+
+### Fortalezas GEO
+
+- Contenido único y específico (ISTQB, Performance Testing, All Pairs)
+- Plataforma freemium con herramientas gratuitas
+- Comunidad activa con ranking
+- Contenido en español (mercado sub-atendido)
+
+### Debilidades GEO
+
+- Sin llms.txt
+- robots.txt básico
+- Sin esquemas JSON-LD
+- Contenido no optimizado para citabilidad
+
+## Plan de Acción
+
+### Fase 1: Configuración Técnica (Semana 1) ✅ COMPLETADO
+
+| Acción                                 | Estado | Archivo                |
+| -------------------------------------- | ------ | ---------------------- |
+| Crear llms.txt                         | ✅     | `/public/llms.txt`     |
+| Actualizar robots.txt para AI crawlers | ✅     | `/public/robots.txt`   |
+| Crear esquemas JSON-LD                 | ✅     | `/public/schema/`      |
+| Crear contenido GEO-optimized          | ✅     | `/public/geo-content/` |
+
+### Fase 2: Optimización de Contenido (Semana 2)
+
+| Acción                               | Prioridad | Herramienta                          |
+| ------------------------------------ | --------- | ------------------------------------ |
+| Crear landing page ISTQB GEO         | Alta      | `/labs/istqb`                        |
+| Crear landing page Performance GEO   | Alta      | `/labs/performance`                  |
+| Crear landing page All Pairs GEO     | Media     | `/labs/allpairs`                     |
+| Crear landing page API Testing GEO   | Media     | `/labs/api-testing-fundamentals`     |
+| Crear landing page SQL/Databases GEO | Alta      | `/assessments/database-fundamentals` |
+
+### Fase 3: Integración Técnica (Semana 3)
+
+| Acción                       | Prioridad | Descripción                            |
+| ---------------------------- | --------- | -------------------------------------- |
+| Agregar meta tags GEO        | Alta      | Optimizar title, description, keywords |
+| Integrar esquemas en páginas | Alta      | JSON-LD en layout.tsx                  |
+| Actualizar sitemap.ts        | Media     | Incluir nuevas páginas GEO             |
+| Crear script de auditoría    | Baja      | `scripts/geo-audit.ts`                 |
+
+### Fase 4: Distribución (Semana 4+)
+
+| Plataforma | Acción                           | Frecuencia |
+| ---------- | -------------------------------- | ---------- |
+| YouTube    | Videos tutoriales ISTQB          | Semanal    |
+| Reddit     | Participar en r/QualityAssurance | 2x/semana  |
+| LinkedIn   | Artículos thought leadership     | Semanal    |
+| Dev.to     | Artículos técnicos               | Quincenal  |
+| Quora      | Responder preguntas QA           | Semanal    |
+
+## Métricas de Éxito
+
+### Corto Plazo (1-3 meses)
+
+- [ ] AIQUAA aparece en respuestas de ChatGPT/Perplexity
+- [ ] Tráfico de referencia AI: +50%
+- [ ] Indexación en Google AI Overviews
+- [ ] 100+ dominios referidos
+
+### Mediano Plazo (3-6 meses)
+
+- [ ] Top 3 en "simulador ISTQB" (español)
+- [ ] Tráfico orgánico: +100%
+- [ ] 500+ usuarios activos mensuales
+- [ ] Mencionado en 10+ artículos QA
+
+### Largo Plazo (6-12 meses)
+
+- [ ] Referencia #1 en QA para LATAM
+- [ ] 1000+ usuarios activos mensuales
+- [ ] Tráfico AI: 30% del total
+- [ ] 50+ menciones en plataformas QA
+
+## Contenido GEO Creado
+
+### Archivos Generados
+
+1. **`/public/llms.txt`** - Guía para AI crawlers
+2. **`/public/robots.txt`** - Permisos para AI crawlers
+3. **`/public/schema/organization.json`** - Esquema Organization
+4. **`/public/schema/software-qa-tools.json`** - Esquemas SoftwareApplication (7 herramientas)
+5. **`/public/schema/ranking.json`** - Esquema ItemList
+6. **`/public/schema/database-ranking.json`** - Esquema ItemList para SQL
+7. **`/public/schema/blog-post.json`** - Esquema BlogPosting
+8. **`/public/geo-content/istqb-que-es.md`** - Contenido ISTQB GEO
+9. **`/public/geo-content/performance-testing.md`** - Contenido Performance GEO
+10. **`/public/geo-content/all-pairs-testing.md`** - Contenido All Pairs GEO
+11. **`/public/geo-content/api-testing.md`** - Contenido API Testing GEO
+12. **`/public/geo-content/sql-databases.md`** - Contenido SQL/Databases GEO
+13. **`scripts/geo-audit.ts`** - Script de auditoría GEO
+
+## Palabras Clave GEO Prioritarias
+
+### ISTQB (⭐⭐⭐⭐⭐)
+
+- "qué es ISTQB"
+- "certificación ISTQB"
+- "preguntas ISTQB"
+- "simulador ISTQB online"
+- "ISTQB CTFL v4.0"
+
+### Performance Testing (⭐⭐⭐⭐)
+
+- "qué es performance testing"
+- "tipos de pruebas de rendimiento"
+- "examen performance testing"
+- "herramientas performance testing"
+
+### All Pairs (⭐⭐⭐)
+
+- "pruebas pairwise"
+- "diseño de casos de prueba"
+- "generador combinaciones"
+- "all pairs testing"
+
+### API Testing (⭐⭐⭐⭐)
+
+- "cómo testear APIs"
+- "API testing fundamentos"
+- "examen API testing"
+
+### SQL/Databases (⭐⭐⭐⭐)
+
+- "qué es SQL"
+- "bases de datos relacionales"
+- "SQL para QA"
+- "consultas SQL para testing"
+- "evaluación SQL online"
+
+## Recomendaciones Adicionales
+
+### Contenido Recomendado
+
+1. **Blog posts** optimizados para citabilidad
+2. **Videos YouTube** con definiciones claras
+3. **Guías PDF** descargables
+4. **Herramientas interactivas** con datos específicos
+
+### Distribución
+
+1. **Reddit** - Participar en comunidades QA
+2. **LinkedIn** - Artículos de thought leadership
+3. **YouTube** - Tutoriales y explicaciones
+4. **Dev.to** - Artículos técnicos
+
+### Monitoreo
+
+1. **Google Search Console** - Tráfico orgánico
+2. **Analytics** - Tráfico de referencia AI
+3. **Perplexity** - Mencionas en IA
+4. **Rank tracking** - Posiciones en búsquedas
+
+## Conclusión
+
+AIQUAA tiene todo lo necesario para dominar el nicho QA en LATAM mediante GEO. La implementación de este plan mejorará significativamente la visibilidad en motores de búsqueda AI, atrayendo más usuarios y estableciendo a AIQUAA como la referencia en QA para la región.
+
+**Próximos pasos inmediatos:**
+
+1. Deploy de los archivos creados
+2. Verificar funcionamiento
+3. Monitorear resultados
+4. Iterar y mejorar
